Seperated out new Message component from chatroom page
